requirements/base.txt
Django==1.9
django-braces==1.8.1
MySQL-python==1.2.5
Pillow==3.0.0
six==1.10.0
Unipath==1.1
wheel==0.24.0
